Physics update receive delta time in seconds now; Remove some conversion warnings; Using static function for calling DebugRenderer essential functions;

+++ ZooidEngine/Animation/Animation.cpp
@@ -32,7 +32,7 @@ namespace ZE
 		Int32 cIndex, nextIndex;
 
 		float portion = _localTime / _duration;
-		cIndex = portion * m_frameCount;
+		cIndex = static_cast<Int32>(portion * m_frameCount);
 		nextIndex = (cIndex + 1) % m_frameCount;
 		float fraction = (portion * m_frameCount) - cIndex;
